+++ .storybook/preview-body.html
@@ -1,4 +1,4 @@
 <script>
-  // For styling purposes
+  // for styling radix components
   document.body.setAttribute("data-ws-component", "Body");
 </script>
