+++ apps/builder/package.json
@@ -13,8 +13,6 @@
     "typecheck": "tsc --noEmit --emitDeclarationOnly false",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "checks": "pnpm typecheck && pnpm test",
-    "storybook:dev": "storybook dev -p 6006",
-    "storybook:build": "storybook build",
     "ci:migrate": "migrations migrate --force",
     "middleware:link": "pnpx vercel link",
     "middleware:env": "pnpx vercel env pull --environment=preview ./.env.preview",
@@ -127,17 +125,12 @@
     "@hono/node-server": "^1.2.0",
     "@jest/globals": "^29.7.0",
     "@remix-run/dev": "^2.9.1",
-    "@storybook/addon-actions": "^8.1.2",
-    "@storybook/addon-essentials": "^8.1.2",
-    "@storybook/addon-links": "^8.1.2",
-    "@storybook/react": "^8.1.2",
     "@testing-library/react-hooks": "^8.0.1",
     "@types/cookie": "^0.6.0",
     "@types/prismjs": "^1.26.0",
     "@types/react": "^18.2.70",
     "@types/react-dom": "^18.2.25",
     "@webstudio-is/jest-config": "workspace:*",
-    "@webstudio-is/storybook-config": "workspace:*",
     "@webstudio-is/tsconfig": "workspace:*",
     "dotenv": "^16.3.1",
     "html-tags": "^3.3.1",
