+++ apps/builder/vite.config.ts
@@ -3,14 +3,11 @@ import { defineConfig } from "vite";
 import { vitePlugin as remix } from "@remix-run/dev";
 import { vercelPreset } from "@vercel/remix/vite";
 
-const isStorybook = process.argv[1]?.includes("storybook") ?? false;
-
 export default defineConfig(({ mode }) => ({
   plugins: [
-    isStorybook === false &&
-      remix({
-        presets: [vercelPreset()],
-      }),
+    remix({
+      presets: [vercelPreset()],
+    }),
   ],
   resolve: {
     conditions: ["webstudio", "import", "module", "browser", "default"],
