+++ package.json
@@ -10,7 +10,7 @@
     "dev": "pnpm --filter='@webstudio-is/builder' --filter='@webstudio-is/prisma-client' --parallel --recursive dev",
     "lint": "eslint . --ext .ts,.tsx --max-warnings 0",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
-    "storybook": "pnpm --parallel --recursive storybook:dev",
+    "storybook:dev": "storybook dev -p 6006",
     "storybook:build": "storybook build",
     "migrations": "cd apps/builder && pnpm migrations --dev",
     "build-figma-tokens": "cd packages/design-system && pnpm build-figma-tokens",
@@ -24,8 +24,18 @@
     "pre-commit": "./node_modules/.bin/nano-staged"
   },
   "devDependencies": {
+    "@fontsource-variable/inter": "^5.0.13",
+    "@fontsource-variable/manrope": "^5.0.15",
+    "@fontsource/roboto-mono": "^5.0.13",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@storybook/addon-actions": "^8.1.2",
+    "@storybook/addon-essentials": "^8.1.2",
+    "@storybook/addon-links": "^8.1.2",
     "@storybook/react": "^8.1.2",
+    "@storybook/react-vite": "^8.1.2",
     "@types/css-tree": "^2.3.1",
+    "@types/node": "^20.12.7",
+    "@types/react": "^18.2.70",
     "@webstudio-is/eslint-config-custom": "workspace:*",
     "esbuild": "^0.20.2",
     "jest": "^29.7.0",
