+++ packages/icons/package.json
@@ -11,19 +11,13 @@
     "dts": "tsc --declarationDir lib/types",
     "generate": "rm -fr src/__generated__ && NODE_OPTIONS='--import=tsx' svgo-jsx svgo-jsx.config.ts && tsx svg-string.ts && prettier --write src/__generated__",
     "typecheck": "tsc",
-    "checks": "pnpm typecheck",
-    "storybook:dev": "storybook dev -p 6006",
-    "storybook:build": "storybook build"
+    "checks": "pnpm typecheck"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^8.1.2",
-    "@storybook/addon-links": "^8.1.2",
-    "@storybook/react": "^8.1.2",
     "@svgo/jsx": "^0.4.2",
     "@types/node": "^20.12.7",
     "@types/react": "^18.2.70",
     "@types/react-dom": "^18.2.25",
-    "@webstudio-is/storybook-config": "workspace:*",
     "@webstudio-is/tsconfig": "workspace:*",
     "react": "18.3.0-canary-14898b6a9-20240318",
     "react-dom": "18.3.0-canary-14898b6a9-20240318",
