--- packages/image/.storybook/main.ts
@@ -1,1 +1,0 @@
-export { default } from "@webstudio-is/storybook-config";
