+++ packages/image/package.json
@@ -11,9 +11,7 @@
     "checks": "pnpm typecheck && pnpm test",
     "dev": "rm -rf lib && esbuild 'src/**/*.ts' 'src/**/*.tsx' --outdir=lib --watch",
     "build": "rm -rf lib && esbuild src/index.ts --outdir=lib --bundle --format=esm --packages=external",
-    "dts": "tsc --project tsconfig.dts.json",
-    "storybook:dev": "storybook dev -p 6006",
-    "storybook:build": "storybook build"
+    "dts": "tsc --project tsconfig.dts.json"
   },
   "dependencies": {
     "react": "18.3.0-canary-14898b6a9-20240318",
@@ -21,13 +19,9 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
-    "@storybook/addon-essentials": "^8.1.2",
-    "@storybook/addon-links": "^8.1.2",
-    "@storybook/react": "^8.1.2",
     "@types/react": "^18.2.70",
     "@types/react-dom": "^18.2.25",
     "@webstudio-is/jest-config": "workspace:*",
-    "@webstudio-is/storybook-config": "workspace:*",
     "@webstudio-is/tsconfig": "workspace:*",
     "react": "18.3.0-canary-14898b6a9-20240318",
     "react-dom": "18.3.0-canary-14898b6a9-20240318",
