--- packages/sdk-components-react/.storybook/main.ts
@@ -1,1 +1,0 @@
-export { default } from "@webstudio-is/storybook-config";
