--- packages/storybook-config/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@webstudio-is/storybook-config",
-  "private": true,
-  "version": "0.0.0",
-  "license": "AGPL-3.0-or-later",
-  "main": "index.ts",
-  "dependencies": {
-    "@fontsource-variable/inter": "^5.0.13",
-    "@fontsource-variable/manrope": "^5.0.15",
-    "@fontsource/roboto-mono": "^5.0.13",
-    "@storybook/addon-essentials": "^8.1.2",
-    "@storybook/addon-links": "^8.1.2",
-    "@storybook/react": "^8.1.2",
-    "@storybook/react-vite": "^8.1.2",
-    "storybook": "^8.1.2",
-    "vite": "^5.2.8"
-  }
-}
