--- packages/storybook-config/setup-fonts.js
@@ -1,9 +1,0 @@
-/* eslint-disable import/no-internal-modules */
-
-// This file is meant to be imported in .storybook/previews.js like so:
-// import "@webstudio-is/storybook-config/setup-fonts";
-
-// this adds <style> tags to the <head> of the document
-import "@fontsource-variable/inter";
-import "@fontsource-variable/manrope";
-import "@fontsource/roboto-mono";
